Rename package poseidon_graphs -> poseidon_graphprep
Follow-up rename of the module (src/poseidon_graphs -> src/poseidon_graphprep)
and distribution (poseidon-graphs -> poseidon-graphprep). Upstream attribution
URLs (github.com/mllam/weather-model-graphs, ...) and CHANGELOG stay unchanged.
