+++ README.md
@@ -1,8 +1,8 @@
-# poseidon-graphs
+# poseidon-graphprep
 
 [![tests](https://github.com/mllam/weather-model-graphs/actions/workflows/ci-tests.yml/badge.svg)](https://github.com/mllam/weather-model-graphs/actions/workflows/ci-tests.yml) [![linting](https://github.com/mllam/weather-model-graphs/actions/workflows/pre-commit.yml/badge.svg)](https://github.com/mllam/weather-model-graphs/actions/workflows/pre-commit.yml) [![Jupyter Book Badge](https://jupyterbook.org/badge.svg)](https://mllam.github.io/weather-model-graphs)
 
-`poseidon-graphs` is a package for creating, visualising and storing graphs used in message-passing graph-based data-driven weather models.
+`poseidon-graphprep` is a package for creating, visualising and storing graphs used in message-passing graph-based data-driven weather models.
 
 The package is designed to use `networkx.DiGraph` objects as the primary data structure for the graph representation right until the graph is to be stored on disk into a specific format.
 This makes the graph generation process modular (every step outputs a `networkx.DiGraph`), easy to debug (visualise the graph at any step) and allows output to different file-formats and file-structures to be easily implemented. More details are given in the [background](https://mllam.github.io/weather-model-graphs/background.html) and [design](https://mllam.github.io/weather-model-graphs/design.html) section of the online [documentation](https://mllam.github.io/weather-model-graphs/).
@@ -10,34 +10,34 @@ This makes the graph generation process modular (every step outputs a `networkx.
 
 ## Installation
 
-If you simply want to install and use `poseidon-graphs` as-is you can install the most recent release directly from pypi with pip
+If you simply want to install and use `poseidon-graphprep` as-is you can install the most recent release directly from pypi with pip
 
 ```bash
-python -m pip install poseidon-graphs
+python -m pip install poseidon-graphprep
 ```
 
 If you want to be able to save to pytorch-geometric data-structure used in
 [neural-lam](https://github.com/mllam/neural-lam) then you will need to install
 pytorch and pytorch-geometric too. This can be done by with the `pytorch`
-optional extra in `poseidon-graphs`:
+optional extra in `poseidon-graphprep`:
 
 ```bash
-python -m pip install poseidon-graphs[pytorch]
+python -m pip install poseidon-graphprep[pytorch]
 ```
 
 This will install the CPU version of pytorch by default. If you want to install
 a GPU variant you should [install that
 first](https://pytorch.org/get-started/locally/) before installing
-`poseidon-graphs`.
+`poseidon-graphprep`.
 
 
-## Developing `poseidon-graphs`
+## Developing `poseidon-graphprep`
 
-The easiest way to work on developing `poseidon-graphs` is to fork the [main repo](https://github.com/mllam/weather-model-graphs) under your github account, clone this repo locally, install [pdm](https://pdm-project.org/en/latest/), create a venv with pdm and then install `poseidon-graphs` (and all development dependencies):
+The easiest way to work on developing `poseidon-graphprep` is to fork the [main repo](https://github.com/mllam/weather-model-graphs) under your github account, clone this repo locally, install [pdm](https://pdm-project.org/en/latest/), create a venv with pdm and then install `poseidon-graphprep` (and all development dependencies):
 
 ```
-git clone https://github.com/<your-github-username>/poseidon-graphs
-cd poseidon-graphs
+git clone https://github.com/<your-github-username>/poseidon-graphprep
+cd poseidon-graphprep
 pdm venv create
 pdm use --venv in-project
 pdm install --dev
@@ -69,13 +69,13 @@ pdm install --group pytorch
 
 # Usage
 
-The best way to understand how to use `poseidon-graphs` is to look at the [documentation](https://mllam.github.io/weather-model-graphs) (which are executable Jupyter notebooks!), to have look at the tests in [tests/](tests/) or simply to read through the source code.
+The best way to understand how to use `poseidon-graphprep` is to look at the [documentation](https://mllam.github.io/weather-model-graphs) (which are executable Jupyter notebooks!), to have look at the tests in [tests/](tests/) or simply to read through the source code.
 
 ## Example, Keisler 2021 flat graph architecture
 
 ```python
 import numpy as np
-import poseidon_graphs as wmg
+import poseidon_graphprep as wmg
 
 # define your (x,y) grid coodinates
 xs, ys = np.meshgrid(np.linspace(0, 1, 32), np.linspace(0, 1, 32))
@@ -97,4 +97,4 @@ for component, graph in graph_components.items():
 
 # Documentation
 
-The documentation is built using [Jupyter Book](https://jupyterbook.org/intro.html) and can be found at [https://mllam.github.io/weather-model-graphs](https://mllam.github.io/weather-model-graphs). This includes background on graph-based weather models, the design principles of `poseidon-graphs` and how to use it to create your own graph architectures.
+The documentation is built using [Jupyter Book](https://jupyterbook.org/intro.html) and can be found at [https://mllam.github.io/weather-model-graphs](https://mllam.github.io/weather-model-graphs). This includes background on graph-based weather models, the design principles of `poseidon-graphprep` and how to use it to create your own graph architectures.
