+++ src/poseidon_graphprep/__init__.py
@@ -1,7 +1,7 @@
 import importlib.metadata
 
 try:
-    __version__ = importlib.metadata.version("poseidon-graphs")
+    __version__ = importlib.metadata.version("poseidon-graphprep")
 except importlib.metadata.PackageNotFoundError:
     __version__ = "unknown"
