+++ tests/test_graph_creation.py
@@ -7,7 +7,7 @@ import numpy as np
 import pytest
 
 import tests.utils as test_utils
-import poseidon_graphs as wmg
+import poseidon_graphprep as wmg
 
 
 def test_create_single_level_mesh_graph():
