+++ tests/test_graph_plots.py
@@ -5,7 +5,7 @@ import numpy as np
 import pytest
 
 import tests.utils as test_utils
-import poseidon_graphs as wmg
+import poseidon_graphprep as wmg
 
 
 def test_plot():
